requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
django-taggit==4.0.0
feedparser==6.0.10
idna==3.4
joblib==1.3.2
nltk==3.8.1
numpy==1.24.4
psycopg2==2.9.7
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
sgmllib3k==1.0.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.0.4
